(image) Add Dockerized GitLint for CI
This is to get ready to set up a GitHub Workflow across
the Ryhino org.

This builds a GitLint image based on the official Python image
and contains some very basic GitLint configs.

It will check all commits from the current HEAD up until `origin/main`.
You will need to mount the repo to check on the host to `/github-workspace/`.